Move database retrieval out and add retry logic
This moves out the database collection for the
influxdb_database resource out to its own function
and does a retry logic if the response from the
InfluxDB HTTP API does not contain the "values"
field in the first series.

This is because when we are bootstrapping InfluxDB
the HTTP API does not add the values field if there
is no databases present, during initialization we
need to wait for the "_internal" database to be created.

This effectively adds plenty of time for InfluxDB to
create the _internal database.